+++ docs/source/release-notes.rst
@@ -1,6 +1,21 @@
 Release Notes
 =============
 
+Version 0.24.2
+______________
+- Engine:
+
+  - Support for MockFile.read() on apps testing hooks
+
+- Plugins:
+
+  - dataframes (beta):
+
+    - Fix: construction from series compatible with pandas
+
+    - Fix: clean index name on new dataframe creation
+
+
 Version 0.24.1
 ______________
 - Plugin: